backend/requirements.txt
fastapi>=0.115,<0.120
uvicorn[standard]>=0.32
pydantic>=2.10
pydantic-settings>=2.6
python-multipart>=0.0.12
google-genai>=0.6
pillow>=11.0
numpy>=2.0
httpx>=0.27
python-json-logger>=2.0
pyyaml>=6.0
pytest>=8.3
pytest-asyncio>=0.24
respx>=0.21
pyjwt[crypto]>=2.9
cryptography>=43.0
cbor2>=5.6
